+++ src/main/java/roomescape/controller/user/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.user;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
