+++ src/main/java/roomescape/controller/view/ViewController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
