+++ src/test/java/roomescape/controller/admin/AdminReservationTimeControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.admin;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
