+++ src/test/java/roomescape/controller/view/ViewControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.view;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
